chore: move everything into one initialiser of okctl
