+++ pkg/credentials/aws/scrape/scrape.go
